Crate wasmer_clif_backend

The Wasmer Cranelift Backend crate is used to compile wasm binary code via parse events from the Wasmer runtime common parser code into machine code.

Constants

VERSION
The current version of this crate

Type Definitions

CraneliftCompiler
Streaming compiler implementation for the Cranelift backed. Compiles web assembly binary into machine code.