Module wasmer_clif_backend::relocation

The relocation package provide two structures: RelocSink, TrapSink. This structures are used by Cranelift when compiling functions to mark any other calls that this function is doing, so we can "patch" the function addrs in runtime with the functions we need.

Modules

call_names

Structs

ExternalRelocation

LocalRelocation

LocalTrapSink

RelocSink
Implementation of a relocation sink that just saves all the information for later

TrapData

TrapSink
Simple implementation of a TrapSink that saves the info for later.

Enums

LibCall

Reloc

RelocationType
Specify the type of relocation

TrapCode

VmCall

VmCallKind

Constants

_IMPL_DESERIALIZE_FOR_ExternalRelocation

_IMPL_DESERIALIZE_FOR_LibCall

_IMPL_DESERIALIZE_FOR_Reloc

_IMPL_DESERIALIZE_FOR_RelocationType

_IMPL_DESERIALIZE_FOR_TrapCode

_IMPL_DESERIALIZE_FOR_TrapData

_IMPL_DESERIALIZE_FOR_TrapSink

_IMPL_DESERIALIZE_FOR_VmCall

_IMPL_DESERIALIZE_FOR_VmCallKind

_IMPL_SERIALIZE_FOR_ExternalRelocation

_IMPL_SERIALIZE_FOR_LibCall

_IMPL_SERIALIZE_FOR_Reloc

_IMPL_SERIALIZE_FOR_RelocationType

_IMPL_SERIALIZE_FOR_TrapCode

_IMPL_SERIALIZE_FOR_TrapData

_IMPL_SERIALIZE_FOR_TrapSink

_IMPL_SERIALIZE_FOR_VmCall

_IMPL_SERIALIZE_FOR_VmCallKind