Module wasmer_clif_backend::signal::unix

Installing signal handlers allows us to handle traps and out-of-bounds memory accesses that occur when running WebAssembly.

This code is inspired by: https://github.com/pepyakin/wasmtime/commit/625a2b6c0815b21996e111da51b9664feb174622

When a WebAssembly module triggers any traps, we perform recovery here.

This module uses TLS (thread-local storage) to track recovery information. Since the four signals we're handling are very special, the async signal unsafety of Rust's TLS implementation generally does not affect the correctness here unless you have memory unsafety elsewhere in your code.

Constants

CAUGHT_ADDRESSES

CURRENT_EXECUTABLE_BUFFER

SETJMP_BUFFER

SETJMP_BUFFER_LEN

Statics

SIGHANDLER_INIT

Functions

call_protected

do_unwind⚠
Unwinds to last protected_call.

get_faulting_addr_and_ip⚠

install_sighandler⚠

longjmp⚠

setjmp⚠

signal_trap_handler

trigger_trap⚠